Add files via upload
Uploaded the first working version. The code is not clean, there are many scars from different tests, but the cleaning process will be done only for the final version. The approach worked when tested with the likelihood operating in image space, with only gaussian noise added to a sample image obtained from an hydrodynamic simulation. Next steps will be: (1) moving to Fourier Space and (2) applying a sampling to emulate an interferometric observation.