Update gan_inference.py
More tests have been conducted and the algorithm responded well also with images not present in the GAN training dataset. The last test done before this commit was done on a map coming from TNG simulations. Next steps will be cleaning up the code and then applying a sampling in Fourier space.